pyproject.toml
[project]
name = "accessibility-assistant-rag"
version = "0.2.0"
description = "RAG assistant per videogiochi accessibili — risposte in italiano, con ChromaDB e Ollama"
readme = "README_NEW.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Luigi Santini", email = "luigi@example.com"}]
keywords = ["rag", "accessibility", "gaming", "ollama", "chromadb"]

dependencies = [
  "chromadb>=0.4.0",
  "ollama>=0.0.11",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "pytest-mock>=3.10.0",
  "black>=23.0.0",
  "pylint>=3.0.0",
  "mypy>=1.0.0",
  "flake8>=6.0.0",
  "isort>=5.0.0",
  "sphinx>=6.0.0",
  "sphinx-rtd-theme>=1.0.0",
  "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/DriftDeV/Accessibility-Assistant"
Documentation = "https://github.com/DriftDeV/Accessibility-Assistant/blob/main/ARCHITECTURE.md"
Repository = "https://github.com/DriftDeV/Accessibility-Assistant.git"
Issues = "https://github.com/DriftDeV/Accessibility-Assistant/issues"

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.poetry]
name = "accessibility-assistant-rag"
version = "0.2.0"
description = "RAG assistant per videogiochi accessibili con Ollama"
authors = ["Luigi Santini <luigi@example.com>"]
license = "MIT"
readme = "README_NEW.md"
repository = "https://github.com/DriftDeV/Accessibility-Assistant"
keywords = ["rag", "accessibility", "gaming", "ollama"]

packages = [
  {include = "accessibility_rag.py"},
  {include = "config.py"},
]

[tool.poetry.dependencies]
python = "^3.10"
chromadb = ">=0.4.0"
ollama = ">=0.0.11"
python-dotenv = ">=1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"
pytest-mock = "^3.10"
black = "^23.0"
pylint = "^3.0"
mypy = "^1.0"
flake8 = "^6.0"
isort = "^5.0"

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | \.eggs
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
  "slow: marks tests as slow",
  "unit: marks tests as unit tests",
  "integration: marks tests as integration tests",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
